+++ libs/naas-abi-core/naas_abi_core/services/triple_store/adaptors/secondary/ApacheJenaTDB2.py
@@ -305,12 +305,14 @@ class ApacheJenaTDB2(ITripleStorePort):
     def query_view(self, view: str, query: str) -> Any:
         return self.query(query)
 
-    def get_subject_graph(self, subject: URIRef) -> Graph:
+    def get_subject_graph(self, subject: URIRef, graph_name: str | URIRef) -> Graph:
         query = f"""
-        CONSTRUCT {{ <{str(subject)}> ?p ?o }}
-        WHERE {{ <{str(subject)}> ?p ?o }}
+        CONSTRUCT {{ <{str(subject)}> ?p ?o . }}
+        WHERE {{ 
+            GRAPH <{str(graph_name)}> 
+            {{ <{str(subject)}> ?p ?o . }} 
+        }}
         """
-
         result = self.query(query)
         if isinstance(result, Graph):
             return result
@@ -321,11 +323,8 @@ class ApacheJenaTDB2(ITripleStorePort):
         assert isinstance(graph_name, URIRef)
         self.query(f"CREATE GRAPH <{str(graph_name)}>")
 
-    def clear_graph(self, graph_name: URIRef | None = None) -> None:
-        if graph_name is None:
-            self.query("CLEAR DEFAULT")
-            return
-
+    def clear_graph(self, graph_name: URIRef) -> None:
+        assert graph_name is not None
         assert isinstance(graph_name, URIRef)
         self.query(f"CLEAR GRAPH <{str(graph_name)}>")